Create ab_testing.py
The purpose of A/B testing is to compare different content and design examples with pairwise comparisons. The purpose of this comparison may differ depending on the system you are A/B testing. You can perform different A/B tests for the homepage, membership page, sales page, landing page or newsletter and follow the change of your targeted action (signing up, returning to sales, clicking..).